+++ robotidy/app.py
@@ -7,15 +7,24 @@ from typing import Dict, List, Optional, Pattern, Tuple
 
 try:
     import rich_click as click
-except ImportError:
+except ImportError:  # Fails on vendored-in LSP plugin
     import click
+
+    escape = None
 
 from robot.api import get_model
 from robot.errors import DataError
 
 from robotidy.config import Config
 from robotidy.disablers import RegisterDisablers
-from robotidy.utils import ModelWriter, StatementLinesCollector, decorate_diff_with_color
+from robotidy.files import get_paths
+from robotidy.transformers import load_transformers
+from robotidy.utils import ModelWriter, StatementLinesCollector, decorate_diff_with_color, escape_rich_markup
+
+try:
+    from robotidy.rich_console import console
+except ImportError:  # Fails on vendored-in LSP plugin
+    console = None
 
 
 class Robotidy:
@@ -105,5 +114,6 @@ class Robotidy:
         if self.config.color:
             output = decorate_diff_with_color(lines)
         else:
-            output = "".join(lines)
-        click.echo(output.encode("ascii", "ignore").decode("ascii"), color=self.config.color)
+            output = escape_rich_markup(lines)
+        for line in output:
+            console.print(line, end="", highlight=False, soft_wrap=True)