+++ CHANGES.md
@@ -22,6 +22,13 @@ Suite Setup         Run Keywords
 
 ## 2.5.0
 
+### Fixes
+- Extended encoding (such as characters like ÁáÉéĚěÍíÓóÚúÝýŮůŘřŤťŠšĎďŽžČčŇň) is now supported in ``--diff`` mode ([#324](https://github.com/MarketSquare/robotframework-tidy/issues/324))
+- ``indent`` parameter is now taken into account by ``NormalizeTags`` transformer ([#323](https://github.com/MarketSquare/robotframework-tidy/issues/323))
+- ``spacecount`` parameter is now taken into account in inline IFs by NormalizeSeparators transformer ([#322](https://github.com/MarketSquare/robotframework-tidy/issues/322))
+
+## 2.5
+
 ### Indent run keywords
 
 New ``IndentNestedKeywords`` transformer for formatting whitespace in run keywords variants such as ``Run Keywords`` or